Gas Leakage Alarm System
There are many incidents like explosions and fire accidents due to LPG gas leakage . Such incidents can cause dangerous effects and also affect the human lives. An IoT based LPG leakage monitoring system will alert the people in the house if a gas leakage will detected by the sensor, also it will send the data value to the Blynk app. So the user can get these data’s at any time. In this project , we are using NodeMCU ESP8266 as a controller. LPG gas sensor is connected with a microcontroller, if a sensor detects the gas it will turn  on the  sensor threshold value.